Modularize Dashboard component (#188)
