+++ gui/src/components/dashboard/capacity-gauge.vue
@@ -52,11 +52,14 @@
 <script lang="ts">
 import { Component, Vue, Prop, Mixins } from "vue-property-decorator";
 import { DiskCapacityDetails } from "./../../models/performance-stats";
-import i18n from "../../i18n";
+import i18n from "./dashboard.json";
 import * as c3 from "c3";
 
 @Component({
-  name: "cortx-capacity-gauge"
+  name: "cortx-capacity-gauge",
+  i18n: {
+    messages: i18n
+  }
 })
 export default class CortxCapacityGauge extends Vue {
   public usedLegendClass = "capacity-used-green";
