+++ gui/src/components/health/cortx-health-view.vue
@@ -16,7 +16,7 @@
 */
 <template>
   <div class="pa-5">
-    <div class="cortx-text-lg cortx-text-bold pr-2">{{ $t("dashboard.healthView") }}</div>
+    <div class="cortx-text-lg cortx-text-bold pr-2">{{ $t("health.healthView") }}</div>
     <div>
       <div class="cortx-health-summary-container">
         <div class="cortx-text-lg cortx-float-l cortx-text-bold">{{componentName}}</div>
@@ -76,7 +76,7 @@
           >
             <span>{{ header.text }}</span>
           </th>
-          <th class="tableheader">{{ $t("dashboard.action") }}</th>
+          <th class="tableheader">{{ $t("health.action") }}</th>
         </tr>
       </template>
       <template v-slot:item="props">
@@ -121,7 +121,7 @@
               @click="$router.push(`/alerts/${props.item.alert_uuid}`)"
             />
            </td>
-           <td v-else>{{ $t("dashboard.NA") }}</td>
+           <td v-else>{{ $t("health.NA") }}</td>
         </tr>
       </template>
     </v-data-table>
